Add confirmation dialog when deleting

+++ src/app/task-list/task-list.component.ts
@@ -1,6 +1,8 @@
 import { Component, OnInit } from '@angular/core';
 import { TaskModel } from '../task.model';
 import { TaskService } from '../task.service';
+import { ConfirmationDialogComponent } from '../confirmation-dialog/confirmation-dialog.component';
+import { MatDialog } from '@angular/material/dialog';
 
 @Component({
   selector: 'app-task-list',
@@ -11,14 +13,24 @@ export class TaskListComponent  {
   tasks: TaskModel[] = [];
   displayedColumns: string[] = ['title', 'description','actions'];
 
-  constructor(private taskService: TaskService) {
+  constructor(private taskService: TaskService ,private dialog: MatDialog) {
     this.taskService.getTasks().subscribe((tasks) => {
       this.tasks = tasks;
     });
 
   }
   deleteTask(task: TaskModel): void {
-    this.taskService.deleteTask(task);
+    const dialogRef = this.dialog.open(ConfirmationDialogComponent, {
+    
+    });
+    dialogRef.afterClosed().subscribe((confirmed: boolean) => {
+      if (confirmed) {
+      console.log('Confirmation received');
+      this.taskService.deleteTask(task);
+    } else {
+      console.log('Confirmation canceled');
+    }
+    });
   }
